knowledge-base/pc-part-enricher/requirements.txt
pandas
pypartpicker2
asyncio
aiohttp
beautifulsoup4
playwright
camoufox[geoip]  